create boardcolumns directory, break components into own files break ticketdialog components into fields and buttons directories clean up files by removing imports, exporting types, deleting unused types move helper functions to utils
