+++ src/components/Kanban/TaskBoard/BoardColumns/ColumnHeader.tsx
@@ -1,0 +1,13 @@
+import { formatTaskStatus } from 'utils/helpers/taskHelpers'
+import { useAppSelector } from 'utils/redux/hooks'
+import { selectVisibleTickets } from 'utils/redux/slices/taskBoardSlice'
+
+export const ColumnHeader = ({ columnStatus }) => {
+  const visibleTickets = useAppSelector(selectVisibleTickets)
+  return (
+    <div className='ticketStatusProgress'>
+      <p>{formatTaskStatus(columnStatus)}</p>
+      <span>{visibleTickets[columnStatus].length}</span>
+    </div>
+  )
+}
