+++ src/components/Kanban/TaskBoard/BoardColumns/ColumnTickets.tsx
@@ -1,0 +1,19 @@
+import { useAppSelector } from 'utils/redux/hooks'
+import { selectVisibleTickets } from 'utils/redux/slices/taskBoardSlice'
+import { TicketTab } from './TicketTab'
+
+export const ColumnTickets = ({ columnStatus, provided }) => {
+  const visibleTickets = useAppSelector(selectVisibleTickets)
+  return (
+    <div
+      className='content'
+      {...provided.droppableProps}
+      ref={provided.innerRef}
+    >
+      {visibleTickets[columnStatus].map((ticketDetail, idx) => (
+        <TicketTab idx={idx} ticketDetail={ticketDetail} />
+      ))}
+      {provided.placeholder}
+    </div>
+  )
+}
