+++ src/components/Kanban/TaskBoard/CreateTicketTab.tsx
@@ -15,6 +15,7 @@ export const CreateTicketTab = ({ columnStatus }) => {
   const userId = useAppSelector(selectUserId)
   const dispatch = useAppDispatch()
   const label = { inputProps: { 'aria-label': 'Checkbox demo' } }
+
   const openCreateTicketDialog = () => {
     dispatch(setVisibleTicketDialog('create'))
     dispatch(
