+++ src/components/Kanban/TaskBoard/TaskBoard.tsx
@@ -2,19 +2,17 @@ import { useEffect, useState } from 'react'
 import '../Ticket.scss'
 import { useAppDispatch, useAppSelector } from 'utils/redux/hooks'
 import { selectProjectTracker } from 'utils/redux/slices/projectSlice'
-import { BoardColumns } from './BoardColumns'
+import { BoardColumns } from './BoardColumns/BoardColumns'
 import { NoTicketsCreated } from './NoTicketsCreated'
 import { BoardHeader } from './BoardHeader'
 import {
-  setVisibleTicketDialog,
   setVisibleTickets,
-  selectVisibleTickets,
   setInitialVisibleTickets,
 } from 'utils/redux/slices/taskBoardSlice'
 import { selectUserId } from 'utils/redux/slices/userSlice'
 import { TicketDialog } from '../TicketDialog/TicketDialog'
 
-// TODO: Rename projectTracker to project
+// TODO: Rename projectTracker to taskBoard
 export const TaskBoard = () => {
   const projectTracker = useAppSelector(selectProjectTracker)
   const userId = useAppSelector(selectUserId)
@@ -45,13 +43,10 @@ export const TaskBoard = () => {
     )
   }, [projectTracker])
 
-  const openDialog = () => dispatch(setVisibleTicketDialog('create'))
-
   return (
     <div className='AllTickets'>
       <BoardHeader />
       {ticketsExist ? <BoardColumns /> : <NoTicketsCreated />}
-      <button onClick={openDialog}>CLICK TO OPEN DIALOG</button>
       <TicketDialog />
     </div>
   )
