+++ src/utils/data/taskBoardConstants.ts
@@ -1,0 +1,29 @@
+import {
+  TaskBoardInterface,
+  TicketFieldsInterface,
+} from 'interfaces/TaskBoardInterface'
+
+export const emptyTicketFields: TicketFieldsInterface = {
+  assignee: '',
+  comments: [],
+  createdBy: '',
+  description: '',
+  dueDate: '',
+  link: '',
+  projectId: '',
+  status: '',
+  title: '',
+}
+
+export const initialTaskBoardState: TaskBoardInterface = {
+  displayAllTickets: true,
+  ticketDialogState: '',
+  ticketFields: emptyTicketFields,
+  visibleTickets: {
+    completed: [],
+    inProgress: [],
+    toDo: [],
+    underReview: [],
+  },
+  visibleTicketDialog: false,
+}
