+++ src/utils/helpers/taskHelpers.ts
@@ -1,8 +1,26 @@
+import { TicketInterface } from 'interfaces'
+import { emptyTicketFields } from 'utils/data/taskBoardConstants'
 import {
-  emptyTicketFields,
   setTicketFields,
   setVisibleTicketDialog,
 } from 'utils/redux/slices/taskBoardSlice'
+
+export const buildTicketPayload = (projectId, userId, ticketFields) => {
+  let ticketPayload: TicketInterface = {
+    projectId,
+    createdBy: userId,
+  }
+
+  // Determine if there is a valid value for Assignee
+  if (!ticketFields.assignee || ticketFields.assignee === 'Unassigned') {
+    const { assignee, ...remainingFields } = ticketFields
+    ticketPayload = { ...remainingFields, ...ticketPayload }
+  } else {
+    ticketPayload = { ...ticketFields, ...ticketPayload }
+  }
+
+  return ticketPayload
+}
 
 export const filterUserTickets = (allTickets, userId) => {
   const filteredTickets = {
@@ -19,6 +37,21 @@ export const filterUserTickets = (allTickets, userId) => {
   }
 
   return filteredTickets
+}
+
+export const formatTaskStatus = (status: string) => {
+  switch (status) {
+    case 'toDo':
+      return 'To Do'
+    case 'inProgress':
+      return 'In Progress'
+    case 'completed':
+      return 'Completed'
+    case 'underReview':
+      return 'Under Review'
+    default:
+      return status
+  }
 }
 
 export const handleCloseVisibleTicketDialog = dispatch => {
