+++ src/utils/redux/slices/taskboardSlice.ts
@@ -1,66 +1,16 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { ProjectTrackerInterface, TaskInterface } from 'interfaces'
+import { ProjectTrackerInterface, TicketInterface } from 'interfaces'
 import { RootState } from '../rootReducer'
 import { filterUserTickets } from 'utils/helpers/taskHelpers'
-import { store } from '../store'
+import { initialTaskBoardState } from 'utils/data/taskBoardConstants'
+import {
+  SetVisibleTicketsReducer,
+  TicketDialogState,
+} from 'interfaces/TaskBoardInterface'
 
-type TicketDialogState = '' | 'create' | 'edit'
-
-export interface TaskBoardInterface {
-  displayAllTickets: boolean
-  ticketDialogState: TicketDialogState
-  ticketFields: TicketFieldsInterface
-  visibleTickets: ProjectTrackerInterface
-  visibleTicketDialog: boolean
-}
-
-export interface TicketFieldsInterface extends TaskInterface {
-  oldStatus?: string
-}
-
-export const emptyTicketFields = {
-  assignee: '',
-  comments: [],
-  createdBy: '',
-  description: '',
-  dueDate: '',
-  image: '',
-  link: '',
-  projectId: '',
-  status: '',
-  title: '',
-}
-
-const initialState: TaskBoardInterface = {
-  displayAllTickets: true,
-  ticketDialogState: '',
-  ticketFields: emptyTicketFields,
-  visibleTickets: {
-    completed: [],
-    inProgress: [],
-    toDo: [],
-    underReview: [],
-  },
-  visibleTicketDialog: false,
-}
-
-export interface SetVisibleTicketsReducer {
-  projectTracker: ProjectTrackerInterface
-  userId?: string
-  changeVisibleTicketType?: boolean
-}
-
-export interface SetVisibleTicketDialogReducer {
-  ticketDialogState: TicketDialogState
-  visibleTicketDialog: boolean
-}
-
-/**
- * Creates a slice for taskBoard with a single reducer to set the image URL.
- */
 const taskBoardSlice = createSlice({
   name: 'taskBoard',
-  initialState,
+  initialState: initialTaskBoardState,
   reducers: {
     setInitialVisibleTickets: (
       state,
@@ -68,22 +18,8 @@ const taskBoardSlice = createSlice({
     ) => {
       state.visibleTickets = action.payload
     },
-    setVisibleTickets: (
-      state,
-      action: PayloadAction<SetVisibleTicketsReducer>
-    ) => {
-      const { changeVisibleTicketType, projectTracker, userId } = action.payload
-      const shouldDisplayAllTickets = changeVisibleTicketType
-        ? !state.displayAllTickets
-        : state.displayAllTickets
-      console.log('shouldDisplayAll', shouldDisplayAllTickets)
-
-      if (shouldDisplayAllTickets) {
-        state.visibleTickets = projectTracker
-      } else {
-        state.visibleTickets = filterUserTickets(projectTracker, userId)
-      }
-      state.displayAllTickets = shouldDisplayAllTickets
+    setTicketFields: (state, action: PayloadAction<TicketInterface>) => {
+      state.ticketFields = { ...state.ticketFields, ...action.payload }
     },
     setVisibleTicketDialog: (
       state,
@@ -102,8 +38,21 @@ const taskBoardSlice = createSlice({
           break
       }
     },
-    setTicketFields: (state, action: PayloadAction<TaskInterface>) => {
-      state.ticketFields = { ...state.ticketFields, ...action.payload }
+    setVisibleTickets: (
+      state,
+      action: PayloadAction<SetVisibleTicketsReducer>
+    ) => {
+      const { changeVisibleTicketType, projectTracker, userId } = action.payload
+      const shouldDisplayAllTickets = changeVisibleTicketType
+        ? !state.displayAllTickets
+        : state.displayAllTickets
+
+      if (shouldDisplayAllTickets) {
+        state.visibleTickets = projectTracker
+      } else {
+        state.visibleTickets = filterUserTickets(projectTracker, userId)
+      }
+      state.displayAllTickets = shouldDisplayAllTickets
     },
   },
 })
